fix test data
