+++ packages/commons/src/client.ts
@@ -16,7 +16,6 @@ export * from './documents'
 export * from './uuid'
 export * from './utils'
 export * from './conditionals/validate'
-export { DEFAULT_ROLES_DEFINITION } from './authentication'
 export * from './fixtures'
 export * from './users'
 export * from './authentication'
